Calvin add responsive delete confirms

+++ resources/js/components/Announcement/SentAnnouncementsIndex.js
@@ -5,7 +5,11 @@ import { Link } from "react-router-dom";
 import Spinner from "../Spinner";
 import {
     Header,
+    Icon
 } from "semantic-ui-react";
+import swal from 'sweetalert'
+
+import '../../../css/SentAnnouncements.css';
 
 class SentAnnouncementsIndex extends React.Component {
     state = {
@@ -38,23 +42,43 @@ class SentAnnouncementsIndex extends React.Component {
     }
 
     handleUnsendToAll = async (id) => {
-        this.setState({ deleting: true })
-        const res = await axios.delete(`${process.env.MIX_API_URL}/announcements/${id}`);
-        if (res.data.status === 200) {
-            this.setState({ deleting: false })
-        }
+        swal({
+            title: "Are you sure?",
+            text: "Once unsent, no one can see the announcement anymore.",
+            icon: "warning",
+            buttons: ["Cancel", "Unsend"],
+            dangerMode: true,
+        }).then(async (willDelete) => {
+            if (willDelete) {
+                this.setState({ deleting: true })
+                const res = await axios.delete(`${process.env.MIX_API_URL}/announcements/${id}`);
+                if (res.data.status === 200) {
+                    this.setState({ deleting: false })
+                }
+            }
+        });
     };
 
     handleManyUnsendToAll = async () => {
-        this.setState({ deleting: true })
-        const { selectedSentAnnouncements } = this.state
-        let selectedSentAnnouncementIds = selectedSentAnnouncements.map(Number);
-        const res = await axios.post(`${process.env.MIX_API_URL}/announcements/deleteMany`, {
-            selectedSentAnnouncementIds: selectedSentAnnouncementIds
+        swal({
+            title: "Are you sure?",
+            text: "Once unsent, no one can see the announcement anymore.",
+            icon: "warning",
+            buttons: ["Cancel", "Unsend"],
+            dangerMode: true,
+        }).then(async (willDelete) => {
+            if (willDelete) {
+                this.setState({ deleting: true })
+                const { selectedSentAnnouncements } = this.state
+                let selectedSentAnnouncementIds = selectedSentAnnouncements.map(Number);
+                const res = await axios.post(`${process.env.MIX_API_URL}/announcements/deleteMany`, {
+                    selectedSentAnnouncementIds: selectedSentAnnouncementIds
+                });
+                if (res.data.status === 200) {
+                    this.setState({ deleting: false })
+                }
+            }
         });
-        if (res.data.status === 200) {
-            this.setState({ deleting: false })
-        }
     }
 
     render() {
@@ -68,8 +92,6 @@ class SentAnnouncementsIndex extends React.Component {
             perPageValues: [5, 10, 20, 25, 100],
             headings: { id: checkAllInput, created_at: 'Created At' },
             sortable: ['name', 'description', 'assignees'],
-            columnsWidth: { actions: "20px", assignees: "60px" },
-            columnsAlign: { id: 'center' },
             requestParametersNames: { query: 'search', direction: 'order' },
             responseAdapter: function (res) {
                 let sentAnnouncementsIDs = res.data.map(a => a.id.toString());
@@ -86,6 +108,10 @@ class SentAnnouncementsIndex extends React.Component {
 
         return (
             <div>
+                <Header as='h2' icon textAlign='center'>
+                    <Icon name='calendar alternate outline' circular />
+                    <Header.Content>Sent Announcements</Header.Content>
+                </Header>
                 <button className="btn btn-danger" onClick={() => { self.handleManyUnsendToAll() }} style={{ marginBottom: "15px" }}>
                     <div style={{ color: "white" }} >
                         <AiFillMinusSquare color="white" size="20" />
@@ -106,19 +132,13 @@ class SentAnnouncementsIndex extends React.Component {
                                                     onChange={self.handleCheckboxTableChange}
                                                     checked={self.state.selectedSentAnnouncements.includes(row.id.toString())} />
                                             );
-                                        case 'assignees':
-                                            return (
-                                                <div style={{ display: "flex", justifyContent: "space-between", width: "60px" }}>
-                                                    {row[column]}
-                                                </div>
-                                            )
                                         case 'description':
                                             return (
                                                 <div dangerouslySetInnerHTML={{ __html: `${row.description}` }} />
                                             )
                                         case 'actions':
                                             return (
-                                                <div style={{ display: "flex", justifyContent: "space-between", width: "20px" }}>
+                                                <div style={{ display: "flex", justifyContent: "start" }}>
                                                     <button className="btn btn-primary" style={{ marginRight: "5px" }}>
                                                         <Link to={'announcements/' + row.id + '/edit/2/sentAnnouncementIndex'}>
                                                             <AiFillEdit color="white" />
