+++ resources/js/components/Customer/CustomerIndex.js
@@ -3,7 +3,13 @@ import ServerTable from 'react-strap-table';
 import { AiFillDelete, AiFillEdit, AiFillPlusSquare, AiFillMinusSquare } from "react-icons/ai";
 import { Link } from "react-router-dom";
 import Spinner from "../Spinner";
+import {
+    Header,
+    Icon
+} from "semantic-ui-react";
+import swal from 'sweetalert'
 
+import '../../../css/Customer.css';
 class CustomerIndex extends React.Component {
     state = {
         selectedCustomers: [],
@@ -34,23 +40,43 @@ class CustomerIndex extends React.Component {
     }
 
     handleDelete = async (id) => {
-        this.setState({ deleting: true })
-        const res = await axios.delete(`${process.env.MIX_API_URL}/customers/${id}`);
-        if (res.data.status === 200) {
-            this.setState({ deleting: false })
-        }
+        swal({
+            title: "Are you sure?",
+            text: "Once deleted, you won't be able to recover the data.",
+            icon: "warning",
+            buttons: ["Cancel", "Delete"],
+            dangerMode: true,
+        }).then(async (willDelete) => {
+            if (willDelete) {
+                this.setState({ deleting: true })
+                const res = await axios.delete(`${process.env.MIX_API_URL}/customers/${id}`);
+                if (res.data.status === 200) {
+                    this.setState({ deleting: false })
+                }
+            }
+        });
     };
 
     handleDeleteMany = async () => {
-        this.setState({ deleting: true })
-        const { selectedCustomers } = this.state
-        let selectedCustomerIds = selectedCustomers.map(Number);
-        const res = await axios.post(`${process.env.MIX_API_URL}/customers/deleteMany`, {
-            selectedCustomerIds: selectedCustomerIds
+        swal({
+            title: "Are you sure?",
+            text: "Once deleted, you won't be able to recover the data.",
+            icon: "warning",
+            buttons: ["Cancel", "Delete"],
+            dangerMode: true,
+        }).then(async (willDelete) => {
+            if (willDelete) {
+                this.setState({ deleting: true })
+                const { selectedCustomers } = this.state
+                let selectedCustomerIds = selectedCustomers.map(Number);
+                const res = await axios.post(`${process.env.MIX_API_URL}/customers/deleteMany`, {
+                    selectedCustomerIds: selectedCustomerIds
+                });
+                if (res.data.status === 200) {
+                    this.setState({ deleting: false })
+                }
+            }
         });
-        if (res.data.status === 200) {
-            this.setState({ deleting: false })
-        }
     }
 
     render() {
@@ -64,7 +90,6 @@ class CustomerIndex extends React.Component {
             perPageValues: [5, 10, 20, 25, 100],
             headings: { id: checkAllInput },
             sortable: ['code', 'name', 'service'],
-            columnsWidth: { id: "5px", actions: "50px" },
             requestParametersNames: { query: 'search', direction: 'order' },
             responseAdapter: function (res) {
                 let customersIDs = res.data.map(a => a.id.toString());
@@ -81,6 +106,10 @@ class CustomerIndex extends React.Component {
 
         return (
             <div>
+                <Header as='h2' icon textAlign='center'>
+                    <Icon name='address book' circular />
+                    <Header.Content>Customers</Header.Content>
+                </Header>
                 <button className="btn btn-primary create" style={{ marginRight: "8px" }}>
                     <Link to={'customers/create'}>
                         <div style={{ color: "white" }} >
@@ -113,7 +142,7 @@ class CustomerIndex extends React.Component {
                                             );
                                         case 'actions':
                                             return (
-                                                <div style={{ display: "flex", justifyContent: "space-between", width: "50px" }}>
+                                                <div style={{ display: "flex", justifyContent: "start" }}>
                                                     <button className="btn btn-primary" style={{ marginRight: "5px" }}>
                                                         <Link to={'customers/' + row.id + '/edit'}>
                                                             <AiFillEdit color="white" />
