+++ resources/js/components/User/UserIndex.js
@@ -3,6 +3,13 @@ import ServerTable from 'react-strap-table';
 import { AiFillDelete, AiFillEdit, AiFillPlusSquare, AiFillMinusSquare } from "react-icons/ai";
 import { Link } from "react-router-dom";
 import Spinner from "../Spinner";
+import {
+    Header,
+    Icon
+} from "semantic-ui-react";
+import swal from 'sweetalert'
+
+import '../../../css/User.css';
 
 class UserIndex extends React.Component {
     state = {
@@ -34,23 +41,43 @@ class UserIndex extends React.Component {
     }
 
     handleDelete = async (id) => {
-        this.setState({ deleting: true })
-        const res = await axios.delete(`${process.env.MIX_API_URL}/users/${id}`);
-        if (res.data.status === 200) {
-            this.setState({ deleting: false })
-        }
+        swal({
+            title: "Are you sure?",
+            text: "Once deleted, you won't be able to recover the data.",
+            icon: "warning",
+            buttons: ["Cancel", "Delete"],
+            dangerMode: true,
+        }).then(async (willDelete) => {
+            if (willDelete) {
+                this.setState({ deleting: true })
+                const res = await axios.delete(`${process.env.MIX_API_URL}/users/${id}`);
+                if (res.data.status === 200) {
+                    this.setState({ deleting: false })
+                }
+            }
+        });
     };
 
     handleDeleteMany = async () => {
-        this.setState({ deleting: true })
-        const { selectedUsers } = this.state
-        let selectedUserIds = selectedUsers.map(Number);
-        const res = await axios.post(`${process.env.MIX_API_URL}/users/deleteMany`, {
-            selectedUserIds: selectedUserIds
+        swal({
+            title: "Are you sure?",
+            text: "Once deleted, you won't be able to recover the data.",
+            icon: "warning",
+            buttons: ["Cancel", "Delete"],
+            dangerMode: true,
+        }).then(async (willDelete) => {
+            if (willDelete) {
+                this.setState({ deleting: true })
+                const { selectedUsers } = this.state
+                let selectedUserIds = selectedUsers.map(Number);
+                const res = await axios.post(`${process.env.MIX_API_URL}/users/deleteMany`, {
+                    selectedUserIds: selectedUserIds
+                });
+                if (res.data.status === 200) {
+                    this.setState({ deleting: false })
+                }
+            }
         });
-        if (res.data.status === 200) {
-            this.setState({ deleting: false })
-        }
     }
 
     render() {
@@ -64,7 +91,6 @@ class UserIndex extends React.Component {
             perPageValues: [5, 10, 20, 25, 100],
             headings: { id: checkAllInput },
             sortable: ['name', 'email', 'role'],
-            columnsWidth: { id: "5px", actions: "30px" },
             requestParametersNames: { query: 'search', direction: 'order' },
             responseAdapter: function (res) {
                 let usersIDs = res.data.map(a => a.id.toString());
@@ -81,6 +107,10 @@ class UserIndex extends React.Component {
 
         return (
             <div>
+                <Header as='h2' icon textAlign='center'>
+                    <Icon name='users' circular />
+                    <Header.Content>Users</Header.Content>
+                </Header>
                 <button className="btn btn-primary create" style={{ marginRight: "8px" }}>
                     <Link to={'users/create'}>
                         <div style={{ color: "white" }} >
@@ -101,7 +131,7 @@ class UserIndex extends React.Component {
                 </button>
                 {
                     deleting ? <Spinner /> :
-                        <ServerTable columns={columns} url={url} options={options} bordered hover updateUrl>
+                        <ServerTable columns={columns} url={url} options={options} bordered hover >
                             {
                                 function (row, column) {
                                     switch (column) {
@@ -113,7 +143,7 @@ class UserIndex extends React.Component {
                                             );
                                         case 'actions':
                                             return (
-                                                <div style={{ display: "flex", justifyContent: "space-between", width: "30px" }}>
+                                                <div style={{ display: "flex", justifyContent: "start" }}>
                                                     <button className="btn btn-primary" style={{ marginRight: "5px" }}>
                                                         <Link to={'users/' + row.id + '/edit'}>
                                                             <AiFillEdit color="white" />
